requirements.txt
fastapi
uvicorn
python-jose
jinja2
passlib
python-multipart
